Minor workflow adjustments (#80)
- Remove macOS runner to save resources (the macOS runner has a [10x
minute
multiplier](https://docs.github.com/en/billing/managing-billing-for-github-actions/about-billing-for-github-actions#minute-multipliers),
so it is more likely to lead to hitting the 2000 minute limit)
   - Testing on ubuntu is reliable enough predictor of macOS support
- Remove `fail-fast` option so that we default back to failing fast if
any platform fails. This is also useful to save resources.
- Start allowing workflow runs on any branch or PR. This will allow for
better developer flow since @ebrelsford and I are soon planning to
target non-main branches in PRs soon as we work to improve the JS
decoder.