Improvements to mvt decoding (#111)
This PR fixes #57, #58, and #75.

It upgrades `no.ecc.vectortile:java-vector-tile` to a version which
supports feature ids but does not break our code (due to
https://github.com/ElectronicChartCentre/java-vector-tile/issues/18 &
https://github.com/maplibre/maplibre-tile-spec/issues/112). This allows
us to drop the `io.github.sebasbaumh:mapbox-vector-tile-java`
dependency. When dropping the
`io.github.sebasbaumh:mapbox-vector-tile-java`, without additional
fixes, the build would fail on missing JTS and Protobuf symbols. This is
because we were depending on specific JTS and Protobuf versions that
`no.ecc.vectortile:java-vector-tile` does not enforce. This is fixed by
adding them to `build.gradle`.