Add justfile with java and js tests (#135)
This will allow CI to run exactly the same tests as developers locally.
This also sets up the framework to compare the output of a directory to
the expected values. Currently, it gets faked by copying the expected
dir to output dir.

Note the `just test-int` -- a good way to compare the results against
the expected values. When called, it:

* runs `clean-int-test` to cleanup the `test/output` dir
* runs `test-run-int` to do any kind of output generation (TODO)
* runs `diff-dirs` to compare the content of `test/output` and
`test/expected`

Another counterpart to it is to run `just bless` -- which does almost
the same, except that it updates the `test/expected` with the new
results.