Fix property decoding in JS implementation (#173)
Only the first 2 properties were being decoded in the JS decoder. This
fixes the problem which was due to a missing implementation of bitsets.
New tests are not yet enabled for JS because other bugs prevent the
tests from passing (refs #169). So, I will address the java side first
and then circle back and enable more property tests in JS once java is
working better.