Add coverage reporting ability to Java tests (#170)
This will automatically generate coverage reports when the tests are
run. They can be found in `build/reports/jacoco/` after a test run.

I'm interested in using these locally to understand which paths in the
java code are utilized (as I work on implementing the JS decoder). In
the future these could be hooked up to CI, but for right now that is a
non-goal.

<img width="1090" alt="Screenshot 2024-06-20 at 12 42 23 PM"
src="https://github.com/maplibre/maplibre-tile-spec/assets/20300/6ae1a855-7fc2-4084-bee3-fab8d14d686a">