Test additional code paths in java decoder (#180)
Currently we do not fully test all these combinations:

 - Vectorized path + advanced encodings
 - Vectorized path + without advanced encodings
 - Non-Vectorized path + advanced encodings
 - Non-Vectorized path + without advanced encodings
 
This PR starts testing all those paths.

The code coverage changes are:

 - **overall:** `36%`to `45%`
 - **mlt.converter:** `64%` to `85%`
 - **mlt.decoder:** `55%` to `71%`

In addition this PR includes:

- Some refactoring in TestUtils.java to share more code between
vectorized and non-vectorized
- Improvements to TestUtils.java around geometry comparisons (refs #178)
 - Improvements to TestUtils.java around property comparisons

Finally, this PR does not attempt to solve any bugs. Rather it finds,
isolates, and writes tests harnesses to capture them in code. Once bugs
are solved in the future it should be easy to update the tests
accordingly.

TODO:
- [x] ~This PR uncovered a number of bugs that do not have open issues
and did not have tests that hit them until now. So I need to create
following issues to capture fixing these.~
     - Added:
        - #181
        - #182 
        - #183 
        - #184
        - #185
        - #186
- [ ] Property nesting in MLT makes comparison to MVT difficult so I
disabled it in this test refactor because comparison to MVT is what the
decoder tests rely on. My sense is that standalone unit tests of nesting
might be more appropriate than including nesting in the decoder tests.
Do others agree?