Rewrite specification for clarity (#525)
using both human and AI, rewrote most of the spec to be more readable,
and flagged some unclear and possibly incorrect moments.

# TODO
- [ ] CPP code uses `Layers` and `FeatureTables` -- I think it will be
easier for everyone to just keep the existing `Layer` naming to simplify
things. In the name of KISS.
- [ ] Current CPP implementation has two integers for each block:
`metadataSize` and `tileDataSize`. Neither seem to be needed because we
agreed to have 1 metadata + 1 feature table for each layer. Since both
are in bytes, it should save considerable space.

---------

Co-authored-by: Frank Elsinga <frank@elsinga.de>
Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>