C++ MLT decoder implementation (#441)
Implements all the necessary types and encodings to load the tiles in
`test/expected`.

Currently builds only with CMake.

Depends on FastPFoR, SIMDE, libfsst, Google test, and (optionally)
nlohmann/json.

When dumped to GeoJSON, results match the output of the nodejs test,
after being modified slightly to produce that output.

---------

Co-authored-by: Bart Louwers <bart@emeel.net>
Co-authored-by: Bart Louwers <bart.louwers@gmail.com>