fix: move extent field back after name (#576)
extent was always right after the name field in each layer, and is still
there in the javascript, but after some recent refactoring it ended up
after the metadata block. I think some geometry decoders might be
affected by it, so it is better to have extent parsed before geometry
parser is instantiated. Moving it back.

### Extras
* use "ALL" for outlines instead of a problematic '*'
* make a simple root justfile recipe to rebuild all expected MLT files