Release from `ts` folder (#602)
This changes the CI to release from the ts folder.
It removes the bump pr version action as I believe it will fails here
with the restriction to create branches as part of this repo.
This means that there will be a need to change the version manually and
open a PR, or do that as part of a different PR...

- I'll test that once #591 is merged