Add Column Mapping to CLI Encoder (#619)
Allows the existing column mapping functionality to be used from the
command line for single-tile or MBTiles conversion.

Multiple mappings can be specified, but they are all applied to all
layers.

Also fixes an issue where the separator was ignored and ':' was always
used. Note that the specified separator is not stored, so it must be
separately provided to the decoder to recover the same property names.

Refactors the argument parsing and try/catch block out of main.