chore(js): use vitest instead of jest (#628)
This moves the tests to use vitest instead of jest.
It moves the test files next to the code as is considered best practice
for typescript/javascript.
It adds the tests to the CI.

Currently, since decode tests fail I added them with "skip" so that they
can be enabled in a different PR.