Refactor column mapping (#625)
This simplifies the process of generating the nested types by pushing
the column-mapping matching logic into `ColumnMapping`, in preparation
for changing the way separators are used and adding explicit mappings.

Re-enables a disabled test relating to mappings. Refactors two tests
into one parameterized test. Adds test coverage of the case where the
base "name" property is encountered after a child "name:en" value.

Updates the CLI encode application to add the ability to apply mappings
to specific Layers/FeatureTables, and add arguments for the
as-yet-unimplemented cases of automatic and explicit mappings.