Bump version to 0.0.1-alpha.12, fix vitest issue (#649)
It seems that a recent change that was made created an issue for vitest
in maplibre-gl repo.
In the future we should probably consider using some bundler like rollup
to bundle the code instead of using `tsc` but if it's not absolutely
necessary I'd like to avoid adding all the relevant rollup
dependencies...

This also bumps the version to 12 in order to create a new release.