Add sources content to source map files (#718)
Source mapping is using the `.ts` files (not inline source mapping).
So in order to get proper source mapping, there's a need to include the
`.ts` files as well in the npm package.
I've discovered this while running vitest with coverage on
maplibre-gl-js as vitest complained that it can't find the source files
linked in the source mapping...
This PR inlines the ts files into the source map files.