fix(js): remove the rest of the static methods and spread operator (#733)
This removes the rest of the static methods.
It also removes spread operator and refactors `StreamMetadata` into a
type and `ZOrderCurve` into a method.
- Related to #725, and build on #730

There are still skipped tests as part of this PR that needs to be
addressed.

I didn't change the file names to keep things simpler to review, but I
would prefer more meaningful file names to the "static" methods.