Update integer decoding tests to use a symmetrical version of the encoder (#740)
This adds some encoding methods to allow symmetrical testing.
I have not written encoder to all decoders, but almost to all of them.
There's a lot of AI involved in writing the encoders I must admit, but I
think the tests are helping to make sure both encoder and decoders are
working as expected.

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
Co-authored-by: Benedikt Vogl <benedikt.vogl@rohde-schwarz.com>